updating for new nbbrowserpdf api
